update-alternatives: fix regression when dropping slave alternatives
When the alternative is changed, it might have to drop a slave alternative
that's no longer provide by the new choice. This must only remove the slave
link if it's actually a link. If it's a file, it must be kept except if
--force is given.

The non-regression tests have been contributed by Andreas Beckmann.

Reported-by: Andreas Beckmann <debian@abeckmann.de>