dpkg: fix possible segfault in findbreakcycle().
The circumstances are not entirely clear because clear_istobes() is
called earlier in the code and should already ensure that clientdata
is allocated for all packages in the database but the stack trace
reported leaves no room for any other interpretation. We must protect
the access to tpkg->clientdata in findbreakcycle() with
ensure_package_clientdata(tpkg).

Probably that some other parts of the code might create new packages in the
in-memory database depending on some specific conditions. It might be that
those conditions only hold for a multiarch-enabled dpkg for example if
the code looks up a package entry for an alternative architecture and
would thus create the package on the fly.

This is pure speculation because I did not push the investigations that
far. It might be something entirely different but it doesn't matter much
because the proposed fix is the same and just ensures that we respect
our API by protecting the access to clientdata.

See https://bugs.launchpad.net/ubuntu/+source/dpkg/+bug/733414 for
details.