Add build-indep and build-arch targets as aliases for build
The current build target is responsible for the generation of
architecture dependent and independent files. The architecture
independent files are just sed'ed and as such they take insignificant
build time, so splitting that from the general build process does not
seem worth the trouble or complexity. Renaming the current build
target to build-arch though would be incorrect.