debian/rules: introduce a maintainer build mode
Adding "maintainer-build" to DEB_BUILD_OPTS changes some compilation
settings to make the build more useful for maintainers. For instance
it enables the silent rules so that it's easier to spot the warnings.

It would be nice to also enable -Werror but that requires further
investigation since enabling this option results in ./configure not
working any more due to warnings generated by several of the
test programs compiled by ./configure itself.