Dpkg::Source::Package: replace register_autopatch() with register_patch()
While register_autopatch() is only able to register a patch as the
automatic patch, register_patch() can register a patch under any
desired patch name.

Also it doesn't not drop the input patch file, leaving that responsibility
to whoever called it. However if the input patch file is empty, it will
remove the target patch from the debian source package.