dpkg-source: simplify message pointing to dpkg-source --commit
The supplementary arguments suggested are awkward and are generally not
needed unless your package is huge and takes a long time to uncompress
and diff.

The best advice is thus “dpkg-source --commit” without any parameter.
Maintainers of big packages will read the manual package and learn
by themselves how to reuse the generated patch.