libdpkg: Use a generic pointer to pass the parsed object data around
This will allow code not parsing binary package control files to be
able to easily reuse the parser, by supplying their own object
structure.

Unfortunately this sacrifices the type-safety of the parse_stanza()
and parse_field() functions, but that only affects one variable,
which should be an acceptable tradeoff given the gain in code reuse.