dpkg: correctly handle a hardlink pointing to a conffile
Normal files tagged as a conffile are not tagged with
fnnf_deferred_rename and due to this the hardlinking code
was assuming that the target is in its final location
already. This is obviously not the case for a conffile.

Fix the code so that the hardlinking logic takes into account
the fact that conffiles are unpacked in .dpkg-new and stay
there until the configuration phase.

Reported-by: Igor Pashev <pashev.igor@gmail.com>