libdpkg: Switch parseversion() to use dpkg_error
This allows to merge back the parseversion_lax() and
version_strict_check() functions into a single parseversion(), which
performs all the validation, but issues different types of dpkg_error
messages on error. Which simplifies parse_db_version() and will allow
for callers to distinguish the error conditions and handle them as
they see fit.