dpkg: Destroy the dpkg_error variables on warnings
The leak on cmpversions() does not pose any issue on normal operation as
the program will exit immediately and the memory will be freed by the
system, but once --command-fd is enabled, it could actually be a problem
due to a long running dpkg process leaking repeatedly.

The leak on parse_db_version() is actually problematic as it can act
repeatedly if there's tons of warnings when parsing.

Regression introduced in commit ccc26399f2f0e295da1206c5ce81fe81a5dc0e21.

Reported-by: Niels Thykier <niels@thykier.net>