build: Use po4a --srcdir and --builddir
This avoids the absolute paths in the po4a Discard output messages,
and simplifies the build infrastructure by not needing the change
directory gymnastics and builddir po4a variable in the po4a.cfg file.

It's been enough time now since #538136 was filed precisely for dpkg
needs, to rely on these “new” options.