libcompat: Update obstack module from gnulib
The version taken is the one before the switch to GPLv3+. With a slight
code revert to not have to include exitfail.c and exitfail.h.

This fixes --disable-nls on systems where obstack support is missing, by
using the local "gettext.h" which honours ENABLE_NLS, instead of just
directly using <libintl.h>.

This also fixes code warnings due to obsolete K&R style prototypes, and
removes the need for bcopy checks.