dpkg-source: do not ignore the automatic patch when checking for unrecorded changes
The former behaviour to always generate the automatic patch required to
not apply it on the tree used as a basis for the patch generation. That
way it could be updated at each build.

With the new behaviour to fail on any new change, we should no longer
ignore it. It must be applied on the reference tree except when
--auto-commit is on.

This should fix 227 (source package) build failures in Debian unstable.