libcompat: Only include libintl.h on --disable-nls if using glibc >= 2
On some systems like MacOS X, gettext's libintl.h redefines setlocale()
to a replacement function from libintl, which will make the build fail
if we end up including the header but not linking to libintl.

This could happen when using --disable-nls to build C++ code and having
the libintl.h header present with the setlocale() replacement function,
but that header is only supposed to be included if building C++ code on
a glibc system.

Restore the logic before commit 8f8e9d9af75672ae051c2d3c11dd797b666b87b8,
but explicitly check that we are not building on a uclibc system.