dpkg: fix trigger setup code to not reset the status of unconfigured packages
The change in commit a36cadba41b19d5908bdfcb658cf01af07c1436a was wrongly
assuming that all packages would be in one of the states
triggers-awaited, triggers-pending or installed when
trig_transitional_activate kicks in.

That is not the case and the status reset should hence only be done when
the current status is already such that the trigger information are
meaningful. transitional_interest_callback_ro() is already taking care
to not add anything in the Triggers-Pending field when the status
is such that it would not make sense.

Reported-by: Hector Oron <hector.oron@gmail.com>