dpkg-source: ignore changes on debian/patches/.dpkg-source-applied
With 2.0 source packages, dpkg-source records as a comment in
debian/patches/.dpkg-source-applied whether the patches have
been applied at build time or at unpack time.

This can lead to differences between the file in the build tree
and the file in the reference tree used to check for changes.
Thus we need to ignore that file in the diff process or the build
might fail considering this difference as a local change that
has not been recorded in a patch.

Reported-by: Jakub Wilk <jwilk@debian.org>