Dpkg::Shlibs: look into multi-arch paths when cross-building
The need for this was discovered when trying to bootstrap a
cross-toolchain against a multiarchified eglibc-source.

We should explicitly add the appropriate multiarch paths to our
library search path. These would be picked up later on anyway in the case
of a native build, but for, e.g., bootstrapping a cross-toolchain the
needed multiarch paths aren't going to be found in ld.so.conf.

Reported-by: Steve Langasek <steve.langasek@linaro.org>