dpkg: Add new deppossi package iterator functions
Since a dependency now points to a pkgset, if you want to find the
correct instance of the package that satisfies the dependency, you
have to take into the account the arch affinity expressed by the
dependency.

The function deppossi_pkg_iter_next() now takes care of this. It
can return multiple packages when the dependency is a wildcard one
(foo:any), you just need to pass the iterator previously created
by deppossi_pkg_iter_new(), and free it with deppossi_pkg_iter_free().

Based-on-patch-by: Raphaël Hertzog <hertzog@debian.org>
Patch-sponsored-by: Linaro Limited

Signed-off-by: Guillem Jover <guillem@debian.org>