libdpkg: Keep Architecture in status entry for non-installed packages
Packages which failed to install or which were removed might still
appear in the status file because we retain the user selection (stored
in pkginfo.pkgwant), and we need to keep the architecture field because
that selection is really arch specific. If we drop the architecture
field, dpkg might confuse this entry with the entry for the native
architecture.

Based-on-patch-by: Raphaël Hertzog <hertzog@debian.org>
Patch-sponsored-by: Linaro Limited

Signed-off-by: Guillem Jover <guillem@debian.org>