dpkg-deb: Treat compression levels independently for each backend
Even if the compression level allowed range supported by gzip/zlib
compatible interfaces is almost the same (0-9), its actual meaning
is backend specific, so treat it that way.

This allows to get native meanings for at least level 0, which maps
to 1 for bzip2, and becomes a new allowed level for xz and lzma.

The reason for the previous behaviour regarding level 0 is historic,
due to zlib treating it that way, and was inadvertently carried over
when new compressors were added to libdpkg.