dpkg-checkbuilddeps: new option -a <arch>
This option tells dpkg-checkbuilddeps to check the build-dependencies
assuming that they must all be satisfied on the given architecture.
In the future build-dependencies might be annotated to indicate whether
we need a package for the host or for the build architecture, but for
now this is a first step that can already be useful.

The normal Multi-Arch logic applies so that so that packages marked as
"Multi-Arch: foreign" can satisfy a build-dependency no matter what
architecture is requested. For the other packages, dpkg-checkbuilddeps
will require a package from the requested architecture.

Signed-off-by: Raphaël Hertzog <hertzog@debian.org>