Stop using brace expansion to install man pages
The reason for using brace expansion in the first place was due to
debhelper starting to abort on empty glob expansion, needed because
translated man pages might not be present if not complete enough.

The installation is now handled by dh_installman which does not abort
on empty glob expansion (and in fairness should have been handling
the installation from the beginning). While the result is clearly a
regression in readability and maintainability it's certainly less ugly
than an executable debhelper config file, it also does not require
bumping debhelper Build-Depends to 8.9.12.