Do not use absolute paths for programs in perl and shell code
The location of programs on the filesystem is not standardized, and as
such using absolute paths is not a portable assumption to make. This
causes test suite errors on at least Mac OS X.

Use just the program name on perl system() calls, when needing the full
path use Dpkg::Path::find_command() in perl code and “which” in shell
code.

Reported-by: Stuart Ambler <stuart@zulazon.com>