build: Reintroduce objdump pattern and tag test case data files
Generating these at check time causes several problems. It requires a
system supporting versioned symbols and a GNU ld command line compatible
linker. Those are not available for example on Mac OS X.

Using libtool could have been an option to portably build the shared
libraries, although that still does not provide versioned symbols where
they are not available. The problem with libtool is that there's no sane
way to refer to the actual built shared library on the build tree.

Introduce a refresh-test-data target to be able to easily regenerate
those files when needed, e.g. when the test suite changes.

Reported-by: Stuart Ambler <stuart@zulazon.com>