dpkg-source: fix issue with relative filename given to --commit
With the “3.0 (quilt)” source format, dpkg-source would happily ignore the
patch-file parameter if it was not found, instead of properly erroring
out. Furthermore a relative filename would also not be found when the
given directory is "." since dpkg-source does a chdir("..") in that case.

Both issues are fixed by this commit.

Closes: #652414
Reported-by: Sam Hartman <hartmans@debian.org>