dpkg-source(1): more explanations for dpkg-source --commit
Clarify the conditions under which it's possible to pass an explicit patch
file to dpkg-source --commit (for "3.0 (quilt)" source packages). This
option has a very narrow purpose: it's mainly to avoid having to
regenerate the patch which can be a costly operation for large packages
and thus assumes several things that one should not forget when using it.