Move copyright information from --version output to file comment headers
Having to keep this information twice is error-prone as it easily gets
out of sync. Having to translate it is bothersome. It's not consistent
across dpkg tools, some do print it some don't. It's currently not
accurate, as the output would need to include the holders for all files
that end up being part of the binary. And listing it in the --versions
output is visually annoying and the wrong place.

Just keep this where it belongs, at the file comment headers, above the
license information.