man: Drop AUTHORS sections in favor of comments in source
This section usually contains copyright notices which can be updated quite
often and thus generate useless work for translators. Furthermore this section
was often confusing by mixing authors of the documented software and authors
of the manual page.

So instead of keeping this in the generated manual page, we move this
information where it belongs, in a file comment header above the license
information.

This commit adds the copyright & license statements to all manual pages in
comments at the top of each file. Information from the git history has been used
where there was no pre-existing copyright statements.