libdpkg: Find the correct package slot in the db for the parsed package
This makes sure to pick the correct slot depending on whether we are
parsing an update log, the status, available or a binary control file.

For the status file we just use the slot matching the architecture.
For update log and control file we pick the slot of the to be installed
package, taking into account architecture cross-grading, and refusing
bogus switches like multiple “Multi-Arch: same” instances into a single
non-“Multi-Arch: same” one. For the available file we use the slot that
will be the candidate for the selection.