dpkg: Use pdb_parse_available in process_archive() on --record-avail
When recording the available information from the binary package, we
don't want it to possibly create a duplicate pkginfo entry with the
same (but mixed) architecture, and having to cleanup the obsolete one.
Instead we ignore the cross-grade slot finding logic, and just use the
old available slot matching the architecture.