libdpkg: Do not reset pkginfo arch_next and set members in pkg_blank()
These are used to link the different instances together in the pkg-db,
and resetting them in pkg_blank() makes using this function not useful
in code accessing the db, as it would unlink the instances from the db.

Now that pkg_blank() does not initialize arch_next, pkgset_blank() has
to do it instead.