dpkg-query: Only load the available file on -l or -W with new --load-avail
Change the default behaviour of not loading the available file depending
on whether arguments have been passed to the commands, to instead adding
a new explicit option to let the user request the additional information
from the available file.

This is needed to avoid confusion given that with multiarch the
available file can contain multiple instances for the same package.