libdpkg: Use arch_empty type to track missing architectures too
The case when the Architecture field is not present is equivalent to
the case of a present Architecture with an empty value, except for
warning purposes. Track missing architectures on the in-core db w/o
overloading arch_none to mean the same, so that we can distinguish
truly unused pkginfo slots.