dpkg: Do not force dependencies on dependtry >= 3, wait until >= 4
To fix the non-working --force-depends-version (bug #57104) in
commit c54fb50e3a08d467955856e81be178b712b7ebdb, a check for
fc_dependsversion was added alongside the existing fc_depends for
dependtry >= 3, as it could not be just replaced because there was
no other code to pickup the fc_depends for that specific case in
dependtry >= 4.

Now that fc_depends always gets applied after all calls to
deppossi_ok_found() have finished, it's safe to remove the misplaced
fc_depends, so that it happens at dependtry >= 4, where it belongs.