dpkg: Allow shared files between Multi-Arch: same packages
If all the already installed packages are synchronized, and the version
to be installed is the same, then we check that the shared file has the
same contents as what's already installed. Otherwise we blindly allow
the updated package to replace the file.

For conffiles, instead of checking against the on-disk version, we
check instead against either:
- the MD5 hash stored for any of the other configured instances of
  the package.
- the MD5 hash of the already present .dpkg-new file for the
  unpacked instances of the packages.

During removal, we only remove the file if it's not shared with other
packages of the same set.

During upgrade, we don't remove the file from other packages containing
the same file if the package is part of the same set. We do this even
for packages which are not “Multi-Arch: same” since there's only one
.list controlg file for the package set and we should not modify it
since it has already been written.

Based-on-patch-by: Raphaël Hertzog <hertzog@debian.org>
Patch-sponsored-by: Linaro Limited

Signed-off-by: Guillem Jover <guillem@debian.org>