build: Always use a local-db
Up until recently the test suite could be run on older dpkg versions,
but as new things get fixed and some behaviour changed, this stops
being the case, a significant example of this is the multiarch tests.
So the test suite should always require the latest dpkg git master.

In addition running on the main database might damage it, for example
if the local dpkg is not multiarch-aware but the main database contains
multiple instances of packages.