s-s-d: Make all pid_is_exec() implementations get a struct stat
This unifies the API across ports, simplifying the code and making it
safer against changes in pid_is_cmd(), and makes it resilient against
symlinks.

This implies removing the generic fallback function that was only being
used on the Hurd, replacing it with an explicit implementation, and
modifying the KVM implementation to match on device and inode numbers
instead of the filenames.