Revert "dpkg-deb: Add epoch when needed to built package names"
This reverts commit a354992c67411ed10499786012701faa33ed8565.

There seems to be several tools with bogus assumptions about the format
of binary package filenames, when they should either just be using the
filenames from .changes files, the Filename field from Packages files,
or the arguments from the command-line. Unforuntately all those bogus
tools need to be fixed first before this change can be considered again,
which right now gets in the way of the upcoming dpkg release.

Also the matching change in dpkg-gencontrol was missing in that commit,
which has been...

Reported-by: Sven Joachim <svenjoac@gmx.de>