libdpkg: Simplify f_booleandefno() and f_multiarch() implementations
This unifies the logic to match the rest of the dumping functions, by
printing an optional field name, the value itself and an optional
trailing newline. These two functions are somewhat special though,
and the reason for the previous code layout, because they do not
print the value if it is null and no field name has been requested.