libdpkg: Do not return from 0 sized buffer_copy() w/o doing filtering
This changes the function to handle the case of a caller specified 0 size
in the same way as if the size requested was dynamic (-1), and there were
0 bytes read, i.e. perform at least the initialization and shutdown of
the filter so that it can compute any required value, if needed.

As an optimization, given that we do not return early from the function
anymore, do not allocate the buffer when the buffer size is 0.

This comes up on 0 sized extracted files, which ended up w/o a computed
hash.