Dpkg::Source::Package: Do not obscure ‘require’ errors with custom errors
When coming out of eval blocks and reporting on errors, make sure $@ is
included as part of the textual output so that the real underlying error
is reported.

  dpkg-source: error: source package format `3.0 (native)' is not supported
  (Perl module Dpkg::Source::Package::V3::native is required)

Of course, that file is being required just fine, the problem turned out
to be missing File::Temp, but without including $@, it was not obvious
due to the custom error message completely hiding this detail.

Improved-by: Jonathan Nieder <jrnieder@gmail.com>
Signed-off-by: Guillem Jover <guillem@debian.org>