Add Build-Depends-Arch and Build-Conflicts-Arch
dpkg currently supports
  Build-Depends (arch all and any)
  Build-Depends-Indep (arch all)
and the same Build-Conflicts.

This patch adds
  Build-Depends-Arch (arch any)
and Build-Conflicts-Arch.

This makes the support for build dependencies for arch all and any
packages symmetrical and more logical. It also permits arch-any
dependencies to be omitted for arch-all-only builds.

This change will not break any existing packages: if the arch any
deps are in Build-Depends, they may be installed unnecessarily,
but they will still be installed. This change just allows an
additional optimisation, and tidies up a slight omission in the
original implementation of build dependencies, and would give the
buildds additional options when arch-all autobuilding is enabled
(especially now it's available as a separate independent arch).

Closes: #629480

Signed-off-by: Roger Leigh <rleigh@debian.org>
Signed-off-by: Guillem Jover <guillem@debian.org>