dpkg: Defer trigger processing if the packages do not fulfill dependencies
The spec specified this behavior, but the code never abided by it.
Change it now that most of the packages with trigger cycles have been
switched to noawaiting triggers. The rest will make dpkg bail out on
upgrade.

We might need to add appropriate versioned Breaks before the Debian
release to get a smoother distribution upgrade.

Closes: #671711