dpkg: Activate all path components for file triggers
File triggers have been activated up to now explicitly whenever seen,
and only the requested pathname. While unpacking or removing, this is
not an issue as dpkg will traverse the hierarchy and trigger parent
directories during the process.

Because conffiles get a two staged installation, first unpacked into
<conffile>.dpkg-new and then installed in place on configure, a trigger
activated only after unpack will not see the <conffile> at its final
location or see it at all if it correctly ignores those file extensions.

This is an issue for conffiles or explicit dpkg-trigger file triggers,
as an interest on parent paths will not activate the trigger as those
parents are not traversed. With the subsequent cause of missed updates
because code has not been run on the actual installation of conffiles.

Closes: #675613, #676061, #676062, #676107, #676118, #676122