libdpkg: Fix explicit file trigger activation
The lex parser passes a pointer to a buffer with the text, and as such
if we just keep a copy of the pointer the contents will change over the
parsing process. Instead copy the contents on trk_file_activate_start(),
and use that copy later on.

Regression introduced in commit 4ce011f68f8e499882db638c065008ce4015cd38.

Closes: #676684