u-a: Only warn for now on out of range priorities on --install
There seems to be packages using priorities > INT_MAX, which although
bogus as they were previously overflowing the int used to store them,
that would cause installation failures when upgrading from squeeze.

Turn this into a warning for now and clamp the values, which will be
switched back to an error after wheezy, in dpkg 1.17.x.

Closes: #676874