dpkg-source: 3.0 (quilt): restore quilt backup files after a failed patch
dpkg-source will now clean up after a failed application of a quilt patch.

This required passing “--reject-file=-” to patch to avoid the creation
of *.rej files that would clutter the source package and intercepting
the failure to restore the files from .pc/$patch/.

dpkg-source will also display a message explaining the most likely cause
of failure (patch applying with fuzz).

Closes: #652970