u-a: Print master and slave link names
Change --query output to print the master and slave link names. This
implies switching the Link field to print the alternative link instead
of the alternative name which was passed already as an argument. Add a
new Name field to print the alternative name.

Although, strictly speaking, this is an interface change, the previous
Link field value was wrong and not really useful, because it's already
known by the caller. A check of the whole archive does not reveal any
user of that field, so it seems pretty safe to do this change.

This also modifies the --config output when thre's a single alternative
to print the alternative link.

Closes: #679010