dpkg-query: Add an Architecture column to --list output
Even if the output format is designed to be human readable, and for
machine parseable output «dpkg-query -W» should be used instead, there's
probably scripts in the wild parsing «dpkg-query -l». Thus, the new
column is inserted before the Description column which is unreliable
to parse anyway as its value will always contain spaces, which are the
column separators, and as such should be pretty safe.

The arch-qualified package names will still be printed, as that
allows to copy-and-paste package names as input to dpkg commands w/o
any possibility of ambiguity.

Closes: #673190

Suggested-by: Jonathan Nieder <jnieder@gmail.com>