dpkg-shlibdeps: Do not use undefined values returned from deps_parse()
When there's a bogus Build-Depends field on the source control file,
and the binary package being built is using a shared library package
that has a Build-Depends-Package field in the symbols file, the
dpkg-shlibdeps code will try to use an undefined value as deps_parse()
failed but the error was not handled before. Instead, error out early.

Closes: #640676

Based-on-patch-by: Raphaël Hertzog <hertzog@debian.org>
Signed-off-by: Guillem Jover <guillem@debian.org>