dpkg: Do not translate SE Linux context to human readable form
The SE Linux context in computer readable form get translated to human
readable form if the mcstransd daemon is running. But if the daemon
fails inbetween then dpkg might try to write wrong data to disk. To
avoid the flakines implied in relying on the state of that daemon,
just never translate the context.

As a side effect this should incur in a slight speed up.

[guillem@debian.org:
 - Coding style fixes.
 - Added code comment. ]

Closes: #679641

Signed-off-by: Guillem Jover <guillem@debian.org>