dpkg: Add new --verify command
For now this checks that the files have not been modified by computing
and comparing their md5sum hashes against the ones in the database.

The only currently supported output format will try to mimick the one
from «rpm --verify», but the default might change in the future in case
it is found cumbersome or lacking, and as such programs parsing the
command output should be explicit about the format they want.

Closes: #187019