Dpkg::Source::Quilt: ensure the target directory exist before copying files
When we restore files in the pop operation, the target directory is
usually pre-existing, but in the case where the patch is removing all the
files in the directory, the usage of patch's --remove-empty-files lead to
the removal of the parent directories which are empty.

This caused a regression in dpkg-source's handling of "3.0 (quilt)" source
package since we switched "3.0 (quilt)" to restore the quilt backup files
to unapply patches. Formerly we were using "patch -R" which was creating
the required directories by itself.

Reported-by: Thomas Koch <thomas@koch.ro>
Closes: #683547