dpkg: Fix logic on reinstallation of previously configured shared conffiles
When checking for the validity of the Conffiles database field, do not
just assume that a package that is just in stat_unpacked has never
before been configured. Instead check if the last configured version
is the same as the currently extracted version, which will mean the
values in the Conffiles field are valid.

Closes: #684776