build: Fix u-a test suite to behave correctly on non-Debian binary paths
The test suite got fixed to not hardcode binary paths, so that it
can work on other systems than Debian, where those binaries might
be located in a different place. Dynamically generate the expected
database file to match the file system layout.

Known to be affecting Gentoo and Mac OS X.