pkg: Change directory to / before executing maintainer scripts
This should give maintainer scripts a more stable and sane environment.

This change does not work too well with relative admindir paths, as they
need to be normalized on input. But then several dpkg tools do not play
nice with relative amindirs anyway, and this needs a general cleanup
which will be done on subsequent commits.

Closes: #686782