dpkg-deb: Add new --deb-format option and base --new and --old on it
The named options are not future-proof, as they will not allow to easily
create packages with specific formats once a future new version gets
introduced. The new option will also allow to request specific minor
revision of a major format, for testing purposes for example.

This also makes the code more explicit as it now refers to specific
format versions when building it.