dpkg: Remove redundant check for otherpkg being pkg on file replaces
This check is already covered by comparing if they are part of the same
package set. As the fileslistvalid check needs to be moved just before
the write_filelist_except() call to be able to insert actions into the
loop that get executed on each iteration, this test there would be too
late by then, and keeping it standalone does not make much sense, so
let's just remove it right away to make the refactoring smoother.