dpkg: Reload the SELinux label database if it has changed
Rework the code to use the selabel API instead of the matchpathcon
wrappers, so that we can query if the label database has changed
during the previous unpack, for example while upgrading the SELinux
policy package. This also allows us to more easily free the resources
allocated for the label database.